* made size of unsigned short explicit and used uint8_t instead * made size of unsigned long int explicit and used uint64_t instead * adjusted header inclusion order: C headers first, then C++, then user-defined headers * using "struct" for classes without member functions * replaced C-style casts by reinterpret_cast<>(...) * made constructors with 1 parameter explicit * applied astyle * added Google's C++ style checker (doc/cpplint.py) * added test cases to Makefile

+++ lola2/src/Place.cc
@@ -12,10 +12,10 @@ in its role as a node, ar contained in Node.*
 
 // Dimensions.h has a type index_type that has at least max(nr of places,nr of transitions) elements
 
+#include <cstdlib>
 #include "Dimensions.h"
 #include "Place.h"
 #include "Node.h"
-#include <cstdlib>
 
 /*!
 
@@ -26,22 +26,23 @@ in its role as a node, ar contained in Node.*
 /// aufräumen für valgrind
 void deletePlaces()
 {
-	free(Place::Hash);
-	free(Place::Capacity);
-	free(Place::CardBits);
-	for(index_type i=0;i<Node::Card[PL];i++)
-	{
-		free(Place::Disabled[i]);
-	}
-	free(Place::Disabled);
-	free(Place::CardDisabled);
+    free(Place::Hash);
+    free(Place::Capacity);
+    free(Place::CardBits);
+    for (index_type i = 0; i < Node::Card[PL]; i++)
+    {
+        free(Place::Disabled[i]);
+    }
+    free(Place::Disabled);
+    free(Place::CardDisabled);
 }
 
 /// Berechne CardBits aus Capacity
-short int Capacity2Bits(capacity_type cap)
+uint8_t Capacity2Bits(capacity_type cap)
 {
-	short int k = 0;
-   	while (cap) {
+    uint8_t k = 0;
+    while (cap)
+    {
         k++;
         cap = cap >> 1;
     }
@@ -50,8 +51,8 @@ short int Capacity2Bits(capacity_type cap)
 
 
 index_type Place::CardSignificant = 0;
-unsigned int * Place::Hash = NULL;
-capacity_type * Place::Capacity = NULL;
-unsigned short int *  Place::CardBits = NULL;
-index_type * Place::CardDisabled = NULL;
-index_type ** Place::Disabled = NULL;
+unsigned int* Place::Hash = NULL;
+capacity_type* Place::Capacity = NULL;
+uint8_t*   Place::CardBits = NULL;
+index_type* Place::CardDisabled = NULL;
+index_type** Place::Disabled = NULL;
